requirements.txt
alembic==1.8.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.3
pandas==1.4.2
psycopg2
scipy==1.8.0
seaborn==0.11.2
SQLAlchemy==1.4.39
sqlparse==0.4.2
uvicorn==0.18.2
python-decouple
pip install 'apache-airflow[amazon]'